+++ service/settlement/src/main/java/jabaclass/settlement/infrastructure/batch/config/SettlementScheduler.java
@@ -1,4 +1,4 @@
-package jabaclass.settlement.infrastructure.batch;
+package jabaclass.settlement.infrastructure.batch.config;
 
 import java.time.YearMonth;
