+++ service/settlement/src/main/java/jabaclass/settlement/infrastructure/batch/dto/MonthlySettlementBatchItem.java
@@ -1,4 +1,4 @@
-package jabaclass.settlement.infrastructure.batch;
+package jabaclass.settlement.infrastructure.batch.dto;
 
 import java.util.List;
