+++ service/settlement/src/main/java/jabaclass/settlement/infrastructure/batch/dto/SettlementTargetCalculationBatchItem.java
@@ -1,4 +1,4 @@
-package jabaclass.settlement.infrastructure.batch;
+package jabaclass.settlement.infrastructure.batch.dto;
 
 import jabaclass.settlement.domain.model.settlement.SettlementTarget;
 import jabaclass.settlement.domain.model.settlement.SettlementTargetCalculation;
